♻️ refactor : 유틸 로직 분리 및 주석 추가 #218
